# examples/weather/services/poller/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: The Poller Service API
    description: US weather forecast poller service API
    version: "1.0"
servers:
    - url: http://localhost:80
      description: Default server for Forecast Poller Service API
paths:
    /poller/location:
        post:
            tags:
                - Poller
            summary: add_location Poller
            description: Adds a location to the polling list.
            operationId: Poller#add_location
            parameters:
                - name: city
                  in: query
                  description: City
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: string
                    description: City
                    example: Santa Barbara
                  example: Santa Barbara
                - name: state
                  in: query
                  description: State
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: string
                    description: State
                    example: CA
                    minLength: 2
                    maxLength: 2
                  example: CA
            responses:
                "202":
                    description: Accepted response.
                    content:
                        application/json:
                            schema:
                                type: string
                                format: binary
                "409":
                    description: 'location_exists: Location already exists'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /poller/subscribe:
        get:
            tags:
                - Poller
            summary: subscribe Poller
            description: Subscribes to weather forecast notifications for a location.
            operationId: Poller#subscribe
            parameters:
                - name: city
                  in: query
                  description: City
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: string
                    description: City
                    example: Santa Barbara
                  example: Santa Barbara
                - name: state
                  in: query
                  description: State
                  allowEmptyValue: true
                  required: true
                  schema:
                    type: string
                    description: State
                    example: CA
                    minLength: 2
                    maxLength: 2
                  example: CA
            responses:
                "101":
                    description: Switching Protocols response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Forecast'
                            example:
                                location:
                                    city: Santa Barbara
                                    lat: 34.4237458
                                    long: -119.7049146
                                    state: CA
                                periods:
                                    - endTime: "2020-01-01T00:00:00Z"
                                      name: Morning
                                      startTime: "2020-01-01T00:00:00Z"
                                      summary: Clear
                                      temperature: 70
                                      temperatureUnit: F
                                    - endTime: "2020-01-01T00:00:00Z"
                                      name: Morning
                                      startTime: "2020-01-01T00:00:00Z"
                                      summary: Clear
                                      temperature: 70
                                      temperatureUnit: F
    /poller/workers:
        delete:
            tags:
                - Poller
            summary: remove_worker Poller
            description: Removes a worker from the poller worker pool.
            operationId: Poller#remove_worker
            responses:
                "204":
                    description: No Content response.
                "417":
                    description: 'too_few: Only one worker left'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
        post:
            tags:
                - Poller
            summary: add_worker Poller
            description: Adds a worker to the poller worker pool.
            operationId: Poller#add_worker
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Worker'
                            example:
                                created_at: "2020-09-01T12:00:00Z"
                                id: "1"
                                jobs:
                                    - created_at: "2020-09-01T12:00:00Z"
                                      key: "1"
                                      payload:
                                        - 123
                                        - 34
                                        - 115
                                        - 116
                                        - 97
                                        - 116
                                        - 101
                                        - 34
                                        - 58
                                        - 34
                                        - 67
                                        - 65
                                        - 34
                                        - 44
                                        - 34
                                        - 99
                                        - 105
                                        - 116
                                        - 121
                                        - 34
                                        - 58
                                        - 34
                                        - 83
                                        - 97
                                        - 110
                                        - 116
                                        - 97
                                        - 32
                                        - 66
                                        - 97
                                        - 114
                                        - 98
                                        - 97
                                        - 114
                                        - 97
                                        - 34
                                        - 125
                                    - created_at: "2020-09-01T12:00:00Z"
                                      key: "1"
                                      payload:
                                        - 123
                                        - 34
                                        - 115
                                        - 116
                                        - 97
                                        - 116
                                        - 101
                                        - 34
                                        - 58
                                        - 34
                                        - 67
                                        - 65
                                        - 34
                                        - 44
                                        - 34
                                        - 99
                                        - 105
                                        - 116
                                        - 121
                                        - 34
                                        - 58
                                        - 34
                                        - 83
                                        - 97
                                        - 110
                                        - 116
                                        - 97
                                        - 32
                                        - 66
                                        - 97
                                        - 114
                                        - 98
                                        - 97
                                        - 114
                                        - 97
                                        - 34
                                        - 125
                                    - created_at: "2020-09-01T12:00:00Z"
                                      key: "1"
                                      payload:
                                        - 123
                                        - 34
                                        - 115
                                        - 116
                                        - 97
                                        - 116
                                        - 101
                                        - 34
                                        - 58
                                        - 34
                                        - 67
                                        - 65
                                        - 34
                                        - 44
                                        - 34
                                        - 99
                                        - 105
                                        - 116
                                        - 121
                                        - 34
                                        - 58
                                        - 34
                                        - 83
                                        - 97
                                        - 110
                                        - 116
                                        - 97
                                        - 32
                                        - 66
                                        - 97
                                        - 114
                                        - 98
                                        - 97
                                        - 114
                                        - 97
                                        - 34
                                        - 125
                                    - created_at: "2020-09-01T12:00:00Z"
                                      key: "1"
                                      payload:
                                        - 123
                                        - 34
                                        - 115
                                        - 116
                                        - 97
                                        - 116
                                        - 101
                                        - 34
                                        - 58
                                        - 34
                                        - 67
                                        - 65
                                        - 34
                                        - 44
                                        - 34
                                        - 99
                                        - 105
                                        - 116
                                        - 121
                                        - 34
                                        - 58
                                        - 34
                                        - 83
                                        - 97
                                        - 110
                                        - 116
                                        - 97
                                        - 32
                                        - 66
                                        - 97
                                        - 114
                                        - 98
                                        - 97
                                        - 114
                                        - 97
                                        - 34
                                        - 125
components:
    schemas:
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: true
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: true
            description: Location already exists
            example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
        Forecast:
            type: object
            properties:
                location:
                    $ref: '#/components/schemas/Location'
                periods:
                    type: array
                    items:
                        $ref: '#/components/schemas/Period'
                    description: Weather forecast periods
                    example:
                        - endTime: "2020-01-01T00:00:00Z"
                          name: Morning
                          startTime: "2020-01-01T00:00:00Z"
                          summary: Clear
                          temperature: 70
                          temperatureUnit: F
                        - endTime: "2020-01-01T00:00:00Z"
                          name: Morning
                          startTime: "2020-01-01T00:00:00Z"
                          summary: Clear
                          temperature: 70
                          temperatureUnit: F
            example:
                location:
                    city: Santa Barbara
                    lat: 34.4237458
                    long: -119.7049146
                    state: CA
                periods:
                    - endTime: "2020-01-01T00:00:00Z"
                      name: Morning
                      startTime: "2020-01-01T00:00:00Z"
                      summary: Clear
                      temperature: 70
                      temperatureUnit: F
                    - endTime: "2020-01-01T00:00:00Z"
                      name: Morning
                      startTime: "2020-01-01T00:00:00Z"
                      summary: Clear
                      temperature: 70
                      temperatureUnit: F
                    - endTime: "2020-01-01T00:00:00Z"
                      name: Morning
                      startTime: "2020-01-01T00:00:00Z"
                      summary: Clear
                      temperature: 70
                      temperatureUnit: F
                    - endTime: "2020-01-01T00:00:00Z"
                      name: Morning
                      startTime: "2020-01-01T00:00:00Z"
                      summary: Clear
                      temperature: 70
                      temperatureUnit: F
            required:
                - location
                - periods
...
        Location:
            type: object
            properties:
                city:
                    type: string
                    description: City
                    example: Santa Barbara
                lat:
                    type: number
                    description: Latitude
                    example: 34.4237458
                    format: double
                long:
                    type: number
                    description: Longitude
                    example: -119.7049146
                    format: double
                state:
                    type: string
                    description: State
                    example: CA
            description: Geographical location
            example:
                city: Santa Barbara
                lat: 34.4237458
                long: -119.7049146
                state: CA
            required:
                - lat
                - long
                - city
                - state
        Period:
            type: object
            properties:
                endTime:
                    type: string
                    description: End time
                    example: "2020-01-01T00:00:00Z"
                    format: date-time
                name:
                    type: string
                    description: Period name
                    example: Morning
                startTime:
                    type: string
                    description: Start time
                    example: "2020-01-01T00:00:00Z"
                    format: date-time
                summary:
                    type: string
                    description: Summary
                    example: Clear
                temperature:
                    type: integer
                    description: Temperature
                    example: 70
                    format: int64
                temperatureUnit:
                    type: string
                    description: Temperature unit
                    example: F
            description: Weather forecast period
            example:
                endTime: "2020-01-01T00:00:00Z"
                name: Morning
                startTime: "2020-01-01T00:00:00Z"
                summary: Clear
                temperature: 70
                temperatureUnit: F
            required:
                - name
                - startTime
                - endTime
                - temperature
                - temperatureUnit
                - summary
...
        Worker:
            type: object
            properties:
                created_at:
                    type: string
                    description: Creation timestamp
                    example: "2020-09-01T12:00:00Z"
                id:
                    type: string
                    description: Worker ID
                    example: "1"
                jobs:
                    type: array
                    items:
                        $ref: '#/components/schemas/Job'
                    description: Worker poll jobs
                    example:
                        - created_at: "2020-09-01T12:00:00Z"
                          key: "1"
                          payload:
                            - 123
                            - 34
                            - 115
                            - 116
                            - 97
                            - 116
                            - 101
                            - 34
                            - 58
                            - 34
                            - 67
                            - 65
                            - 34
                            - 44
                            - 34
                            - 99
                            - 105
                            - 116
                            - 121
                            - 34
                            - 58
                            - 34
                            - 83
                            - 97
                            - 110
                            - 116
                            - 97
                            - 32
                            - 66
                            - 97
                            - 114
                            - 98
                            - 97
                            - 114
                            - 97
                            - 34
                            - 125
                        - created_at: "2020-09-01T12:00:00Z"
                          key: "1"
                          payload:
                            - 123
                            - 34
                            - 115
                            - 116
                            - 97
                            - 116
                            - 101
                            - 34
                            - 58
                            - 34
                            - 67
                            - 65
                            - 34
                            - 44
                            - 34
                            - 99
                            - 105
                            - 116
                            - 121
                            - 34
                            - 58
                            - 34
                            - 83
                            - 97
                            - 110
                            - 116
                            - 97
                            - 32
                            - 66
                            - 97
                            - 114
                            - 98
                            - 97
                            - 114
                            - 97
                            - 34
                            - 125
            example:
                created_at: "2020-09-01T12:00:00Z"
                id: "1"
                jobs:
                    - created_at: "2020-09-01T12:00:00Z"
                      key: "1"
                      payload:
                        - 123
                        - 34
                        - 115
                        - 116
                        - 97
                        - 116
                        - 101
                        - 34
                        - 58
                        - 34
                        - 67
                        - 65
                        - 34
                        - 44
                        - 34
                        - 99
                        - 105
                        - 116
                        - 121
                        - 34
                        - 58
                        - 34
                        - 83
                        - 97
                        - 110
                        - 116
                        - 97
                        - 32
                        - 66
                        - 97
                        - 114
                        - 98
                        - 97
                        - 114
                        - 97
                        - 34
                        - 125
                    - created_at: "2020-09-01T12:00:00Z"
                      key: "1"
                      payload:
                        - 123
                        - 34
                        - 115
                        - 116
                        - 97
                        - 116
                        - 101
                        - 34
                        - 58
                        - 34
                        - 67
                        - 65
                        - 34
                        - 44
                        - 34
                        - 99
                        - 105
                        - 116
                        - 121
                        - 34
                        - 58
                        - 34
                        - 83
                        - 97
                        - 110
                        - 116
                        - 97
                        - 32
                        - 66
                        - 97
                        - 114
                        - 98
                        - 97
                        - 114
                        - 97
                        - 34
                        - 125
                    - created_at: "2020-09-01T12:00:00Z"
                      key: "1"
                      payload:
                        - 123
                        - 34
                        - 115
                        - 116
                        - 97
                        - 116
                        - 101
                        - 34
                        - 58
                        - 34
                        - 67
                        - 65
                        - 34
                        - 44
                        - 34
                        - 99
                        - 105
                        - 116
                        - 121
                        - 34
                        - 58
                        - 34
                        - 83
                        - 97
                        - 110
                        - 116
                        - 97
                        - 32
                        - 66
                        - 97
                        - 114
                        - 98
                        - 97
                        - 114
                        - 97
                        - 34
                        - 125
                    - created_at: "2020-09-01T12:00:00Z"
                      key: "1"
                      payload:
                        - 123
                        - 34
                        - 115
                        - 116
                        - 97
                        - 116
                        - 101
                        - 34
                        - 58
                        - 34
                        - 67
                        - 65
                        - 34
                        - 44
                        - 34
                        - 99
                        - 105
                        - 116
                        - 121
                        - 34
                        - 58
                        - 34
                        - 83
                        - 97
                        - 110
                        - 116
                        - 97
                        - 32
                        - 66
                        - 97
                        - 114
                        - 98
                        - 97
                        - 114
                        - 97
                        - 34
                        - 125
            required:
                - id
                - jobs
                - created_at
...